% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicpdf.R
\name{biccdf}
\alias{biccdf}
\title{Bivariate conditional CDF}
\usage{
biccdf(object, newdata, cores = 1)
}
\arguments{
\item{object}{an object of class \code{bivinereg}.}

\item{newdata}{data.frame of new response and covariate values for which to
compute the bivariate conditional distribution.}

\item{cores}{integer; the number of cores to use for computations.}
}
\description{
Calculates the bivariate conditional distribution \eqn{C_{V_1 V_2 | \mathbf{U}}}
of the responses \eqn{(V_1, V_2)} given covariates \eqn{\mathbf{U}}.
}
\examples{
# load sample data
data(data)

# fit vine regression model
(fit <- bivinereg(cbind(U1, U4) ~ U2 + U3 + U5 + U6,
                  data = data,
                  family_set = "parametric",
                  selcrit = "bic",
                  uscale = TRUE))

# calculate the bivariate conditional CDF
biccdf(fit, data[1:5,])

}
