% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bichdr.R
\name{bichdr}
\alias{bichdr}
\title{Bivariate conditional highest density region}
\usage{
bichdr(
  object,
  newdata,
  scale = "xscale",
  xlim = c(-5, 5),
  ylim = c(-5, 5),
  ...
)
}
\arguments{
\item{object}{an object of class \code{bivinereg}.}

\item{newdata}{data.frame with one row of covariate values.}

\item{scale}{One of \code{"xscale", "uscale", "zscale"}. The scale on which the HDR should be plotted. \code{scale = "xscale"} only works if \code{object} was fitted with using \code{uscale = FALSE} in \code{\link[=bivinereg]{bivinereg()}}.}

\item{xlim}{Vector of two integers. Interval on the x-axis in which the HDR will be calculated. Automatically set to \code{c(0, 1)} if \code{scale = "uscale"}.}

\item{ylim}{Vector of two integers. Interval on the y-axis in which the HDR will be calculated. Automatically set to \code{c(0, 1)} if \code{scale = "uscale"}.}

\item{...}{Further arguments to \code{\link[=geom_hdr_fun]{geom_hdr_fun()}}.}
}
\value{
A \link{ggplot2} object.
}
\description{
Plot the highest density region (HDR) of the bivariate conditional density for a set of covariate values.
}
\examples{
# load sample data
data(data)

# fit vine regression model
(fit <- bivinereg(cbind(U1, U4) ~ U2 + U3 + U5 + U6,
                  data = data,
                  family_set = "parametric",
                  selcrit = "bic",
                  uscale = TRUE))

# Plot HDR
bichdr(fit, data[1,], scale = "zscale")

}
\seealso{
\code{\link[=bivinereg]{bivinereg()}} \link{ggplot2} \code{\link[=geom_hdr_fun]{geom_hdr_fun()}}
}
