% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicsample.R
\name{bicsample}
\alias{bicsample}
\title{Sample conditionally from a Y-vine regression model}
\usage{
bicsample(object, n, newdata, cores = 1)
}
\arguments{
\item{object}{an object of class \code{bivinereg}.}

\item{n}{integer; the number of samples to generate.}

\item{newdata}{data.frame of covariate values for which to sample.}

\item{cores}{integer; the number of cores to use for computations.}
}
\value{
A data.frame of samples with two columns corresponding to the
responses.
}
\description{
Sample conditionally from a Y-vine regression model
}
\examples{
# load sample data
data(data)

# fit vine regression model
(fit <- bivinereg(cbind(U1, U4) ~ U2 + U3 + U5 + U6,
                  data = data,
                  family_set = "parametric",
                  selcrit = "bic",
                  uscale = TRUE))

# Sample conditionally
bicsample(fit, 10, data[1,])

}
