% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivinereg.R
\name{bivinereg}
\alias{bivinereg}
\title{Y-vine regression models}
\usage{
bivinereg(
  formula,
  data,
  family_set = "parametric",
  selcrit = "aic",
  par_1d = list(),
  weights = numeric(),
  cores = 1,
  ...,
  uscale = FALSE
)
}
\arguments{
\item{formula}{an object of class "formula"; same as \code{\link[=lm]{lm()}}.}

\item{data}{data frame (or object coercible by \code{\link[=as.data.frame]{as.data.frame()}}) containing
the variables in the model.}

\item{family_set}{see \code{family_set} argument of \code{\link[rvinecopulib:bicop]{rvinecopulib::bicop()}}.}

\item{selcrit}{selection criterion based on conditional log-likelihood.
\code{"loglik"} (default) imposes no correction; other choices are
\code{"aic"} and \code{"bic"}.}

\item{par_1d}{list of options passed to \code{\link[kde1d:kde1d]{kde1d::kde1d()}}, must be one value
for each margin, e.g. \code{list(xmin = c(0, 0, NaN))} if the responses have
non-negative support.}

\item{weights}{optional vector of weights for each observation.}

\item{cores}{integer; the number of cores to use for computations.}

\item{...}{further arguments passed to \code{\link[rvinecopulib:bicop]{rvinecopulib::bicop()}}.}

\item{uscale}{if TRUE, bivinereg assumes that marginal distributions have been
taken care of in a preliminary step.}
}
\value{
An object of class bivinereg. It is a list containing the elements
\describe{ \item{formula}{the formula used for the fit.}
\item{selcrit}{criterion used for variable selection.}
\item{model_frame}{the data used to fit the regression model.}
\item{margins}{list of marginal models.}
\item{vine}{an \code{\link[rvinecopulib:vinecop_dist]{rvinecopulib::vinecop_dist()}} object containing the fitted
Y-vine.} \item{stats}{fit statistics such as conditional
log-likelihood/AIC/BIC and p-values for each variable's contribution.}
\item{order}{order of the covariates chosen by the variable selection
algorithm.} \item{selected_vars}{indices of selected variables.} }
\code{summary.bivinereg()} shows the contribution of each selected variable with
the associated p-value derived from a likelihood ratio test.
\code{summary_yvine()} gives a more detailed and flexible summary of the fitted
Y-vine.
}
\description{
Sequential estimation of a regression Y-vine as described in Tepegjozova and
Czado (2022).
}
\examples{
# load sample data
data(data)

# fit vine regression model
(fit <- bivinereg(cbind(U1, U4) ~ U2 + U3 + U5 + U6,
                  data = data,
                  family_set = "parametric",
                  selcrit = "bic",
                  uscale = TRUE))

# inspect model
summary(fit)

}
\references{
Tepegjozova and Czado (2022), Bivariate vine copula based
regression, bivariate level and quantile curves,
https://arxiv.org/abs/2205.02557
}
