% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Data}
\format{
An object of class \code{data.frame} with 500 rows and 6 columns.
}
\usage{
data(data)
}
\description{
Simulated Y-vine data
}
\examples{
data(data)
}
\keyword{datasets}
