% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicpdf.R
\name{mcpdf}
\alias{mcpdf}
\title{Marginal conditional PDF}
\usage{
mcpdf(object, newdata, margin, cores = 1)
}
\arguments{
\item{object}{an object of class \code{bivinereg}.}

\item{newdata}{data.frame of new response and covariate values for which to
compute the marginal conditional density.}

\item{margin}{integer; the margin (1 or 2) for which to calculate the
conditional PDF.}

\item{cores}{integer; the number of cores to use for computations.}
}
\description{
Calculates the marginal conditional density \eqn{c_{V_i | \mathbf{U}}} of a
response \eqn{V_i, i = 1, 2,} given covariates \eqn{\mathbf{U}}.
}
\examples{
# load sample data
data(data)

# fit vine regression model
(fit <- bivinereg(cbind(U1,U4) ~ U2 + U3 + U5 + U6,
                  data = data,
                  family_set = "parametric",
                  selcrit = "bic"))

# calculate marginal conditional PDF of U1 given the covariates
mcpdf(fit, data[1:5,], margin = 1)

}
