% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary_yvine}
\alias{summary_yvine}
\title{Summary for Y-vines}
\usage{
summary_yvine(
  object,
  digits = 2,
  trees = "ALL",
  rows = "ALL",
  names_cols = c("tree", "edge", "cond'd", "conditioning", "family", "rota",
    "parameters", "df", "tau", "loglik"),
  ...
)
}
\arguments{
\item{object}{An object of class "vinecop". The fitted vine from \code{bivinereg()}.}

\item{digits}{Digits in summary.}

\item{trees}{The trees for which to print the summary. Either a vector of
numbers or "ALL" (default) to print all trees.}

\item{rows}{Specific rows of the summary to be printed. Either a vector of
numbers or "ALL" (default) for all rows.}

\item{names_cols}{A vector of names of columns to be printed in the summary.}
}
\description{
A summary for Y-vines tailored to Y-vine regression.
}
