% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicpdf.R
\name{bicprob}
\alias{bicprob}
\title{Bivariate conditional probability}
\usage{
bicprob(object, newdata, v1_min, v1_max, v2_min, v2_max, cores = 1)
}
\arguments{
\item{object}{an object of class \code{bivinereg}.}

\item{newdata}{matrix of covariate values for which to compute
the bivariate conditional probability.}

\item{v1_min}{lower limit of first response.}

\item{v1_max}{upper limit of first response.}

\item{v2_min}{lower limit of second response.}

\item{v2_max}{upper limit of second response.}

\item{cores}{integer; the number of cores to use for computations.}
}
\description{
Calculates the bivariate conditional probability of the responses given the
covariates.
}
\details{
This function double integrates the bivariate conditional density using the
function \code{integral2} from the \code{pracma} package.
}
\examples{
# load sample data
data(data)

# fit vine regression model
(fit <- bivinereg(cbind(U1,U4) ~ U2 + U3 + U5 + U6,
                  data = data,
                  family_set = "parametric",
                  selcrit = "bic"))

# calculate the conditional probability that 0.3 <= U1 <= 0.6,
# 0.3 <= U4 <= 0.6, given the covariates
bicprob(fit, data[1:5,], rep(0.3, 5), rep(0.6, 5), rep(0.3, 5), rep(0.6, 5))

}
\seealso{
\code{\link{bicprob2}}
}
