% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicpdf.R
\name{mccdf}
\alias{mccdf}
\title{Marginal conditional CDF}
\usage{
mccdf(object, newdata, margin, cores = 1)
}
\arguments{
\item{object}{an object of class \code{bivinereg}.}

\item{newdata}{matrix of response and covariate values for which to compute
the marginal conditional distribution.}

\item{margin}{integer; the margin for which to calculate the conditional CDF.}

\item{cores}{integer; the number of cores to use for computations.}
}
\description{
Calculates the marginal conditional distribution of a response given the
covariates and other response.
}
\examples{
# load sample data
data(data)

# fit vine regression model
(fit <- bivinereg(cbind(U1,U4) ~ U2 + U3 + U5 + U6,
                  data = data,
                  family_set = "parametric",
                  selcrit = "bic"))

# calculate the marginal conditional CDF of U1 given U4 and the covariates
mccdf(fit, data[1:5,], margin = 1)

}
