% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicpdf.R
\name{mcpdf}
\alias{mcpdf}
\title{Marginal conditional PDF}
\usage{
mcpdf(object, newdata, margin, cores = 1)
}
\arguments{
\item{object}{an object of class \code{bivinereg}.}

\item{newdata}{matrix of response and covariate values for which to compute
the bivariate conditional density.}

\item{margin}{integer; the margin for which to calculate the conditional PDF.}

\item{cores}{integer; the number of cores to use for computations.}
}
\description{
Calculates the marginal conditional density of a response given the
covariates.
}
\examples{
# load sample data
data(data)

# fit vine regression model
(fit <- bivinereg(cbind(U1,U4) ~ U2 + U3 + U5 + U6,
                  data = data,
                  family_set = "parametric",
                  selcrit = "bic"))

# calculate marginal conditional PDF of U1 given the covariates
mcpdf(fit, data[1:5,], margin = 1)

}
